+++ Samples~/Analytics/AnalyticsCallbackExample.cs
@@ -83,11 +83,8 @@ namespace GameLovers.UiService.Examples
 	/// </summary>
 	public class AnalyticsExample : MonoBehaviour
 	{
-		[SerializeField] private UiConfigs _uiConfigs;
+		[SerializeField] private PrefabRegistryUiConfigs _uiConfigs;
 
-		[Header("Sample Prefabs")]
-		[SerializeField] private GameObject[] _presenterPrefabs;
-		
 		[Header("UI Buttons")]
 		[SerializeField] private Button _loadButton;
 		[SerializeField] private Button _openButton;
@@ -112,12 +109,7 @@ namespace GameLovers.UiService.Examples
 			SubscribeToAnalyticsEvents();
 			
 			// Initialize UI Service with analytics
-			var loader = new PrefabRegistryUiAssetLoader();
-			foreach (var prefab in _presenterPrefabs)
-			{
-				var presenter = prefab.GetComponent<UiPresenter>();
-				loader.RegisterPrefab(presenter.GetType().Name, prefab);
-			}
+			var loader = new PrefabRegistryUiAssetLoader(_uiConfigs);
 
 			_uiService = new UiService(loader, _analytics);
 			_uiService.Init(_uiConfigs);