+++ Samples~/UiSets/UiSetsExample.cs
@@ -23,11 +23,8 @@ namespace GameLovers.UiService.Examples
 			SettingsPanel = 2
 		}
 
-		[SerializeField] private UiConfigs _uiConfigs;
+		[SerializeField] private PrefabRegistryUiConfigs _uiConfigs;
 
-		[Header("Sample Prefabs")]
-		[SerializeField] private GameObject[] _presenterPrefabs;
-		
 		[Header("UI Buttons")]
 		[SerializeField] private Button _loadSetButton;
 		[SerializeField] private Button _openSetButton;
@@ -41,12 +38,7 @@ namespace GameLovers.UiService.Examples
 		private void Start()
 		{
 			// Initialize UI Service
-			var loader = new PrefabRegistryUiAssetLoader();
-			foreach (var prefab in _presenterPrefabs)
-			{
-				var presenter = prefab.GetComponent<UiPresenter>();
-				loader.RegisterPrefab(presenter.GetType().Name, prefab);
-			}
+			var loader = new PrefabRegistryUiAssetLoader(_uiConfigs);
 
 			_uiService = new UiService(loader);
 			_uiService.Init(_uiConfigs);